HANA_HOMEPLANNER_CRAWLER/consumer/requirements.txt
kafka-python==2.0.2
pymongo==4.6.0
requests==2.31.0
google-generativeai==0.3.2
pdfplumber==0.10.3
boto3==1.34.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0